+++ pinot-hadoop/src/main/java/org/apache/pinot/hadoop/job/BaseSegmentJob.java
@@ -32,6 +32,7 @@ import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.pinot.common.Utils;
 import org.apache.pinot.common.config.TableConfig;
+import org.apache.pinot.common.data.Schema;
 import org.apache.pinot.hadoop.utils.PushLocation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -73,16 +74,16 @@ public abstract class BaseSegmentJob extends Configured {
     }
   }
 
-  /**
-   * This method is currently implemented in SegmentCreationJob and SegmentPreprocessingJob due to a dependency on
-   * the hadoop filesystem, which we can only get once the job begins to run.
-   * We return null here to make it clear that for now, all implementations of this method have to support
-   * reading from a schema file. In the future, we hope to deprecate reading the schema from the schema file in favor
-   * of mandating that a schema is pushed to the controller.
-   */
   @Nullable
-  protected org.apache.pinot.common.data.Schema getSchema() throws IOException {
-    return null;
+  protected Schema getSchema()
+      throws IOException {
+    try (ControllerRestApi controllerRestApi = getControllerRestApi()) {
+      if (controllerRestApi != null) {
+        return controllerRestApi.getSchema();
+      } else {
+        throw new RuntimeException("Unable to get schema");
+      }
+    }
   }
 
   /**
